requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.4
Django==3.0.12
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
drfpasswordless==1.5.7
flake8==3.9.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
isort==5.8.0
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.7.0
pyflakes==2.3.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
